chore: move Location.kt from /map to /journey and delete map folder since he is empty
