+++ app/src/androidTest/java/com/android/brewr/ui/e2eTest.kt
@@ -34,7 +34,7 @@ import com.android.brewr.model.journey.CoffeeTaste
 import com.android.brewr.model.journey.Journey
 import com.android.brewr.model.journey.JourneysRepository
 import com.android.brewr.model.journey.ListJourneysViewModel
-import com.android.brewr.model.map.Location
+import com.android.brewr.model.journey.Location
 import com.android.brewr.model.user.UserRepository
 import com.android.brewr.model.user.UserViewModel
 import com.android.brewr.ui.explore.ExploreScreen
