+++ app/src/androidTest/java/com/android/brewr/ui/explore/CoffeeShopInformationCardTest.kt
@@ -11,7 +11,7 @@ import androidx.test.uiautomator.UiSelector
 import com.android.brewr.model.coffee.CoffeeShop
 import com.android.brewr.model.coffee.Hours
 import com.android.brewr.model.coffee.Review
-import com.android.brewr.model.map.Location
+import com.android.brewr.model.journey.Location
 import com.android.brewr.ui.navigation.NavigationActions
 import com.android.brewr.ui.navigation.Screen
 import java.time.LocalDate
