+++ app/src/androidTest/java/com/android/brewr/ui/explore/MapScreenTest.kt
@@ -17,7 +17,7 @@ import com.android.brewr.model.journey.CoffeeTaste
 import com.android.brewr.model.journey.Journey
 import com.android.brewr.model.journey.JourneysRepository
 import com.android.brewr.model.journey.ListJourneysViewModel
-import com.android.brewr.model.map.Location
+import com.android.brewr.model.journey.Location
 import com.google.firebase.Timestamp
 import kotlinx.coroutines.runBlocking
 import org.junit.Assert.*
