+++ app/src/androidTest/java/com/android/brewr/ui/overview/JourneyRecordScreenTest.kt
@@ -13,7 +13,7 @@ import androidx.compose.ui.test.performScrollTo
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.android.brewr.model.coffee.CoffeeShop
 import com.android.brewr.model.journey.*
-import com.android.brewr.model.map.Location
+import com.android.brewr.model.journey.Location
 import com.android.brewr.ui.navigation.NavigationActions
 import com.android.brewr.ui.navigation.Screen
 import com.google.firebase.Timestamp
