+++ app/src/main/java/com/android/brewr/model/journey/JourneysRepositoryFirestore.kt
@@ -4,7 +4,6 @@ import android.util.Log
 import com.android.brewr.model.coffee.CoffeeShop
 import com.android.brewr.model.coffee.Hours
 import com.android.brewr.model.coffee.Review
-import com.android.brewr.model.map.Location
 import com.android.brewr.model.user.User
 import com.google.android.gms.tasks.Task
 import com.google.firebase.Firebase
@@ -71,8 +70,9 @@ class JourneysRepositoryFirestore(
   /**
    * Retrieves all journeys from the Firestore database.
    *
-   * @param onSuccess The callback to call with the list of journeys if the operation is successful.
-   * @param onFailure The callback to call if the operation fails.
+   * @param [onSuccess] The callback to call with the list of journeys if the operation is
+   *   successful.
+   * @param [onFailure] The callback to call if the operation fails.
    */
   /**
    * Retrieves all journeys from the Firestore database in real-time using addSnapshotListener.
