+++ app/src/test/java/com/android/brewr/model/journey/JourneysRepositoryFirestoreTest.kt
@@ -6,7 +6,6 @@ import androidx.test.core.app.ApplicationProvider
 import com.android.brewr.model.coffee.CoffeeShop
 import com.android.brewr.model.coffee.Hours
 import com.android.brewr.model.coffee.Review
-import com.android.brewr.model.map.Location
 import com.android.brewr.model.user.User
 import com.google.android.gms.tasks.OnSuccessListener
 import com.google.android.gms.tasks.Task
