+++ app/src/test/java/com/android/brewr/model/journey/ListJourneysViewModelTest.kt
@@ -3,7 +3,6 @@ package com.android.brewr.model.journey
 import com.android.brewr.model.coffee.CoffeeShop
 import com.android.brewr.model.coffee.Hours
 import com.android.brewr.model.coffee.Review
-import com.android.brewr.model.map.Location
 import com.google.firebase.Timestamp
 import org.hamcrest.CoreMatchers.`is`
 import org.hamcrest.MatcherAssert.assertThat
